requirements.txt
pandas
numpy
requests
openpyxl
Pillow
python-dotenv
gspread
oauth2client
streamlit
plotly
google-generativeai
groq
